Adding verify_status decorator to unlock endpoints
