+++ backend/audit/views/views.py
@@ -15,7 +15,6 @@ from django.http import JsonResponse
 
 from audit.fixtures.excel import FORM_SECTIONS, UNKNOWN_WORKBOOK
 
-
 from audit.forms import (
     AuditorCertificationStep1Form,
     AuditorCertificationStep2Form,
@@ -42,6 +41,7 @@ from audit.validators import (
     validate_auditee_certification_json,
     validate_auditor_certification_json,
 )
+from audit.verify_status import verify_status
 from audit.utils import FORM_SECTION_HANDLERS
 
 from dissemination.remove_workbook_artifacts import remove_workbook_artifacts
@@ -59,38 +59,6 @@ logger = logging.getLogger(__name__)
 
 def _friendly_status(status):
     return dict(SingleAuditChecklist.STATUS_CHOICES)[status]
-
-
-def verify_status(status):
-    """
-    Decorator to be applied to view request methods (i.e. get, post) to verify
-    that the submission is in the correct state before allowing the user to
-    proceed. An incorrect status usually happens via direct URL access. If the
-    given status does not match the submission's, it will redirect them back to
-    the submission progress page.
-    """
-
-    def decorator_verify_status(request_method):
-        def verify(view, request, *args, **kwargs):
-            report_id = kwargs["report_id"]
-
-            try:
-                sac = SingleAuditChecklist.objects.get(report_id=report_id)
-            except SingleAuditChecklist.DoesNotExist:
-                raise PermissionDenied("You do not have access to this audit.")
-
-            # Return to checklist, the Audit is not in the correct state.
-            if sac.submission_status != status:
-                logger.warning(
-                    f"Expected submission status {status} but it's currently {sac.submission_status}"
-                )
-                return redirect(f"/audit/submission-progress/{sac.report_id}")
-            else:
-                return request_method(view, request, *args, **kwargs)
-
-        return verify
-
-    return decorator_verify_status
 
 
 class MySubmissions(LoginRequiredMixin, generic.View):
